+++ Smart/Smart/Pages/Instructors/Grading/Index.cshtml
@@ -1,46 +1,103 @@
 ﻿@page
-@model Smart.Pages.Instructors.Classes.IndexModel
+@model Smart.Pages.Instructors.Grading.IndexModel
 
 @{
-    ViewData["Title"] = "Index";
+    ViewData["Title"] = "Grading";
 }
 
-<h1>Index</h1>
+<h1>Grade Book</h1>
+<form method="get" id="ClassForm">
+    <div class="row">
+        <select id="ClassId" class="form-control" name="classId" asp-items="Model.Classes">
+            <option value="">Select Class</option>
+        </select>
+    </div>
+</form>
+    @*<p>
+            <a asp-page="Create">Add New Grade</a>
+        </p>*@
 
-<p>
-    <a asp-page="Create">Create New</a>
-</p>
-<table class="table">
-    <thead>
-        <tr>
-            <th>
-                @Html.DisplayNameFor(model => model.Class[0].Course)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Class[0].Term)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Class[0].ClassId)
-            </th>
-            <th></th>
-        </tr>
-    </thead>
-    <tbody>
-@foreach (var item in Model.Class) {
-        <tr>
-            <td>
-                @Html.DisplayFor(modelItem => item.Course.Name)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Term.Description)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.ClassId)
-            </td>
-            <td>
-                <a asp-page="./Edit" asp-route-id="@item.ClassId">Select Class</a>
-            </td>
-        </tr>
+    @*<table class="table">
+            <thead>
+                <tr>
+                    <th>
+                        Class ID
+                    </th>
+                    <th>
+                        Assessment Title
+                    </th>
+                    <th>
+                        Description
+                    </th>
+                    <th>
+                        Points Possible
+                    </th>
+                    <th></th>
+                </tr>
+            </thead>
+            <tbody>
+                @foreach (var item in Model.GradingData.Assessments)
+        {
+                    <tr>
+                        <td>
+                            @Html.DisplayFor(modelItem => item.Class.ClassId)
+                        </td>
+                        <td>
+                            @Html.DisplayFor(modelItem => item.Title)
+                        </td>
+                        <td>
+                            @Html.DisplayFor(modelItem => item.Description)
+                        </td>
+                        <td>
+                            @Html.DisplayFor(modelItem => item.PointsPossible)
+                        </td>
+                        <td>
+                            <a asp-page="./Index" asp-route-assessmentId="@item.AssessmentId">View Assessments</a>
+                        </td>
+                    </tr>
+                }
+            </tbody>
+        </table>*@
+
+
+    @if (Model.clsId != 0)
+        {
+            <h3>Student Assessments</h3>
+            <table class="table">
+                <tr>
+                    <th>Student ID</th>
+                    <th>Points</th>
+                    <th>Comments</th>
+                    <th></th>
+                </tr>
+
+                @foreach (var item in Model.GradingData.StudentAssessments)
+        {
+            <tr>
+                <td>
+                    @item.StudentId
+                </td>
+                <td>
+                    @item.PointsAwarded
+                </td>
+                <td>
+                    @item.Comments
+                </td>
+                <td>
+                    <a asp-page="./Edit" asp-route-studentAssessmentID="@item.AssessmentId">Edit Grade</a>
+                </td>
+            </tr>
+                }
+
+            </table>
+        }
+
+@section Scripts{ 
+    <script>
+        $(document).ready(function () {
+            $('#ClassId').change(function () {
+                $('#ClassForm').submit();
+            });
+        });
+    </script>
 }
-    </tbody>
-</table>
