+++ Smart/Smart/Pages/Instructors/Grading/Index.cshtml.cs
@@ -4,11 +4,13 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Smart.Data;
 using Smart.Models;
+using Smart.Models.InstructorViewModels;
 
-namespace Smart.Pages.Instructors.Classes
+namespace Smart.Pages.Instructors.Grading
 {
     public class IndexModel : PageModel
     {
@@ -18,14 +20,45 @@ namespace Smart.Pages.Instructors.Classes
         {
             _context = context;
         }
-
-        public IList<Class> Class { get;set; }
-
-        public async Task OnGetAsync()
+        [BindProperty]
+        public int clsId { get; set; }
+        public GradingIndexData GradingData { get; set; }
+        [BindProperty]
+        public List<SelectListItem> Classes { get; set; }
+        public async Task<IActionResult> OnGetAsync(int? classId)
         {
-            Class = await _context.Class
-                .Include(c => c.Course)
-                .Include(c => c.Term).ToListAsync();
+            GradingData = new GradingIndexData();
+            GradingData.Classes = await _context.Class.ToListAsync();
+            GradingData.StudentAssessments = await _context.StudentAssessment.ToListAsync();
+            GradingData.Assessments = await _context.Assessment.ToListAsync();
+            GradingData.Students = await _context.Student.ToListAsync();
+
+            Classes = await _context.Class.Select(c =>
+                                            new SelectListItem
+                                            {
+                                                Value = c.ClassId.ToString(),
+                                                Text = c.Course.Name.ToString()
+                                            }).ToListAsync();
+            //.Include(a => a.Assessments)
+            //.Include(a => a.Students)
+            //.Include(a => a.StudentAssessments).ToListAsync();
+
+            if (classId != null)
+            {
+                clsId = classId.Value;
+                try
+                {
+                    Assessment assessment = GradingData.Assessments.Where(a => a.ClassId == classId).Single();
+                    GradingData.StudentAssessments = assessment.StudentAssessments.ToList();
+                }
+                catch
+                {
+                    clsId = 0;
+                }
+
+            }
+
+            return Page();
         }
     }
 }
