--- Smart/Smart/Pages/Instructors/Grading/Index2.cshtml
@@ -1,115 +1,0 @@
-﻿@page "{id:int?}"
-@model Smart.Pages.Instructors.Grading.IndexModel
-
-@{
-    ViewData["Title"] = "Grading";
-}
-
-<h1>Grade Book</h1>
-
-<label class="control-label">Select Class</label>
-<select id="selectClasses" class="form-control" name="ClassID" asp-for="ClassID" asp-items="Model.Classes"></select>
-
-@*@if (Model.ClassID != 0)
-{
-    <h1>test</h1>
-}*@
-
-
-@*<p>
-        <a asp-page="Create">Add New Grade</a>
-    </p>*@
-
-@*<table class="table">
-        <thead>
-            <tr>
-                <th>
-                    Class ID
-                </th>
-                <th>
-                    Assessment Title
-                </th>
-                <th>
-                    Description
-                </th>
-                <th>
-                    Points Possible
-                </th>
-                <th></th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach (var item in Model.GradingData.Assessments)
-    {
-                <tr>
-                    <td>
-                        @Html.DisplayFor(modelItem => item.Class.ClassId)
-                    </td>
-                    <td>
-                        @Html.DisplayFor(modelItem => item.Title)
-                    </td>
-                    <td>
-                        @Html.DisplayFor(modelItem => item.Description)
-                    </td>
-                    <td>
-                        @Html.DisplayFor(modelItem => item.PointsPossible)
-                    </td>
-                    <td>
-                        <a asp-page="./Index" asp-route-assessmentId="@item.AssessmentId">View Assessments</a>
-                    </td>
-                </tr>
-            }
-        </tbody>
-    </table>*@
-
-
-@*@if (Model.GradingData.StudentAssessments != null)
-    {
-        <h3>Student Assessments</h3>
-        <table class="table">
-            <tr>
-                <th>Student ID</th>
-                <th>Points</th>
-                <th>Comments</th>
-                <th></th>
-            </tr>
-
-            @foreach (var item in Model.GradingData.StudentAssessments)
-    {
-        <tr>
-            <td>
-                @item.StudentId
-            </td>
-            <td>
-                @item.PointsAwarded
-            </td>
-            <td>
-                @item.Comments
-            </td>
-            <td>
-                <a asp-page="./Edit" asp-route-studentAssessmentID="@item.AssessmentId">Edit Grade</a>
-            </td>
-        </tr>
-            }
-
-        </table>
-    }*@
-
-<script>
-    $(document).ready(function () {
-         $("#btnSelectClass").click(function () {
-             // Get the value from 'Select1'
-             var value = $("#selectClasses").val();
-         });
-
-        $.ajax({
-            url: '@Url.Action("ActionName", "ControllerName")',
-            data: { valueToQuery: $("#selectClasses").val() },
-            success: function (data) {
-               // The data is the result
-            }
-        });
-    });
-
-
-</script>
