+++ Smart/Smart/Pages/Instructors/Grading/Index3.cshtml
@@ -1,0 +1,46 @@
+﻿@page
+@model Smart.Pages.Instructors.Classes.IndexModel
+
+@{
+    ViewData["Title"] = "Index";
+}
+
+<h1>Index</h1>
+
+<p>
+    <a asp-page="Create">Create New</a>
+</p>
+<table class="table">
+    <thead>
+        <tr>
+            <th>
+                @Html.DisplayNameFor(model => model.Class[0].Course)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Class[0].Term)
+            </th>
+            <th>
+                @Html.DisplayNameFor(model => model.Class[0].ClassId)
+            </th>
+            <th></th>
+        </tr>
+    </thead>
+    <tbody>
+@foreach (var item in Model.Class) {
+        <tr>
+            <td>
+                @Html.DisplayFor(modelItem => item.Course.Name)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.Term.Description)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.ClassId)
+            </td>
+            <td>
+                <a asp-page="./Edit" asp-route-id="@item.ClassId">Select Class</a>
+            </td>
+        </tr>
+}
+    </tbody>
+</table>
