+++ Smart/Smart/Pages/Instructors/Grading/Index3.cshtml.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.EntityFrameworkCore;
+using Smart.Data;
+using Smart.Models;
+
+namespace Smart.Pages.Instructors.Classes
+{
+    public class IndexModel : PageModel
+    {
+        private readonly Smart.Data.ApplicationDbContext _context;
+
+        public IndexModel(Smart.Data.ApplicationDbContext context)
+        {
+            _context = context;
+        }
+
+        public IList<Class> Class { get;set; }
+
+        public async Task OnGetAsync()
+        {
+            Class = await _context.Class
+                .Include(c => c.Course)
+                .Include(c => c.Term).ToListAsync();
+        }
+    }
+}
